Add Student Sonar frontend code.
The frontend is architected using the new hotness of ES6 and web components.
There are two custom elements currently in use: x-student card, which provides
a small clickable avatar + name, and x-student-view, which is a full-page view
showing the student's name/school/email/avatar, and contains the different
analytics for each student.

At the moment, we show their participation on Slack using a github-like heatmap
calendar, and a list of their review requests on Review Board. Future
improvements will show code reviews, status reports, and links to their notes.

There's a lot still to be improved here, but this is a good start.

Reviewed at https://reviews.reviewboard.org/r/6897/