Add a status reports section.
This change adds a new section to the student view which links to each of the
status reports. This currently requires me to go through and add anchors to
each status reports blog post, and then put the relevant data into
students.json. That workflow is kind of annoying, but it will suffice for now.

Testing done:
Checked that the status reports showed up correctly.

Reviewed at https://reviews.reviewboard.org/r/6939/