Change the way static media is delivered.
Instead of allowing express to serve any file in the current directory, specify
the exact paths which it should serve. This required making a change in the way
that index.html is served as well (apparently it was being shown by
express.static instead of the route to /).

Testing done:
Verified that everything still worked.

Reviewed at https://reviews.reviewboard.org/r/6938/