Add a new 'summary entry' element, make things pretty.
This change cleans up the sections on the student view with a new element that
wraps a section. In addition to this, I've redone a lot of the styles to make
things pretty.

Testing done:
Loaded all the different students and saw that things worked as expected.

Reviewed at https://reviews.reviewboard.org/r/6940/