On the frontend, show the right thing if there's a ?student= query string.
Because we're using pushState to modify the history, and the HTML/JS that's
sent to the client is the same no matter what, we need to do some routing on
the frontend. This now looks at the query string to see if there's a ?student=,
and if so, opens up the relevant student view.

Testing done:
Opened URLs with ?student= in them and saw it do the right thing.

Reviewed at https://reviews.reviewboard.org/r/6941/