Deploy all the things!
This sets us up with a build system that compiles the frontend HTML, CSS and
JavaScript, as well as makes some changes to the config loading to be able to
pull out configuration from the environment instead of a local file.

This is primarily oriented towards deployment on Heroku, since they have a free
tier that handles what we need, but it should be pretty applicable to any
Node.js hosting.

Testing done:
- Did a bunch of builds and verified that everything was correct.
- Deployed!

Reviewed at https://reviews.reviewboard.org/r/6957/