Alternately load students.json from a URL.
When deploying, we need to have students.json available somewhere. Rather than
committing this to git, I've stuck it into an S3 bucket. We now check to see if
it's present on the filesystem and use that, and otherwise load it from a URL.

Testing done:
Verified that sonar still loads correctly.

Reviewed at https://reviews.reviewboard.org/r/6950/