Add a "Demo Videos" section, refactor links.
This adds a new section to the student view to link to all of our sweet demo
videos. As part of this, I've refactored the links structure to reduce
duplication of code.

Testing done:
Checked all the links and saw that things still worked as expected.

Reviewed at https://reviews.reviewboard.org/r/6955/