Fix building packages of rb-reviews-api
There are a couple things which broke when trying to build an egg for the
reviews API extension.

First, in the resources, there was a circular reference when trying to define
the resource when the file was loaded as part of the static media builder.
Importing reviewboard.webapi.resources and accessing one of the resources first
makes sure that everything is registered properly.

Second, the package name in setup.py has to be a bytestring.

Testing done:
Built an egg successfully.

Reviewed at https://reviews.reviewboard.org/r/6980/