Switch to using templates for rendering.
This change switches us over to using handlebars templates for assembling the
HTML that's rendered to the client, rather than just sending a static HTML
page. This is in preparation for adding a bunch of new functionality and
additional views.

Also part of this is the addition of the errorhandler middleware, and set-up
for including images as part of the distribution and build process.

Testing done:
- Ran a build and verified that the resulting templates were processed
  correctly.
- Verified that things still loaded the same as before this change.

Reviewed at https://reviews.reviewboard.org/r/7226/