Switch data storage over to MongoDB.
As sonar gets more complicated, it's no longer feasible to manually maintain a
big JSON file to store all the data. This change swaps that out in favor of
MongoDB, hosted via a heroku add-on.

The schema itself is currently identical to what I was using in students.json,
so there's very little to change on the front-end. The only difference is that
there's less depth to get to the actual students, since I removed the whole
top-level "term" nonsense.

The nice thing about this is that it will now be very easy to add lots of
functionality that requires the web server to be able to write data (such as
having students submit their status reports directly).

Testing done:
Verified that everything still loaded correctly.

Reviewed at https://reviews.reviewboard.org/r/7235/