Add sessions and login with slack.
This change adds access control to sonar, using Slack OAuth as a login
provider. This uses the 'passport' framework to handle the guts, and the
passport-slack provider to talk to Slack itself. Sessions are stored in
mongodb.

Testing done:
Logged in and out a whole bunch.

Reviewed at https://reviews.reviewboard.org/r/7237/