Add a timeline section to the bottom of the student view.
With the volume of students this term, we've had some difficultly checking in
whether students are doing weekly code drops. This change adds a new section to
the student view that has a timeline of events. At the moment, this shows when
new review requests are posted, and when an update to a review request that
includes a new diff is posted.

As part of this, I ended up refactoring the calendar in the main view to be a
reusable element which automatically sorts its entries in the right order.

Testing done:
Looked at lots of different students and saw the expected timelines.

Reviewed at https://reviews.reviewboard.org/r/7258/