Redesign the sonar experience.
This change is the start of a major redesign of how sonar is put together. I've
brought in bootstrap, which gives us a lot of styles, layout, and various
components for free.

All of the students for the active terms are now listed in a sidebar, along
with other "components" such as the calendar. This gives us a good place to put
all the different features that I plan on building.

Testing done:
Checked calendar and student views.

Reviewed at https://reviews.reviewboard.org/r/7534/