Add a "Manage" mode and calendar editing.
This change adds the beginnings of the administration mode which is available
to mentors. This mode is entered by clicking a "Manage" toggle at the top-right
of the page.

This mode is currently implemented for the Calendar page. Switching into manage
mode will show all of the events (even ones in the past), and some additional
UI. Each event can be edited or deleted, and new events can be added.

Testing done:
Added, edited, and deleted calendar items. Toggled to manage mode a bunch.

Reviewed at https://reviews.reviewboard.org/r/7536/