Fix event handling for clicking on user sidebar entries.
For the sidebar, the event handling was using `ev.target`, which in the case of
clicking directly on a name (rather than next to it) would be the `<span>`
element rather than the `<a>`. Changing to `ev.currentTarget` fixes this.

Testing done:
Clicked around and saw that user links always worked no matter which part of
the sidebar entry I clicked on.

Reviewed at https://reviews.reviewboard.org/r/7551/