Add a page showing the list of available student projects.
This page scrapes a configured Asana project for a list of tasks that students
can work on. This allows projects to maintain all the tasks in Asana and deal
with the assignment and management there, and to reflect those available
projects on a page that can be linked to by non-members.

We currently display the task name, tags, and description. If the task is
assigned to a user with a matching e-mail address, that user's avatar will show
up next to the task name. I'm hoping to make it so we can assign tasks to guest
users in Asana through the API so that students don't have to care about
learning yet another tool (for now).

If configured, the data is stored in memcached for 5 minutes, so that repeated
reloads won't hammer the Asana API too much.

This page is currently only available when logged in, since all of the pages
are. I'm going to try to whip up a change soon that makes it so the main shell
of the app with a welcome page and projects list are available without
authentication, and then logging in gets you the rest of the features.

This is based on Christian's change at https://reviews.reviewboard.org/r/7575/

Testing done:
- Viewed both the API and the page, and saw the student projects.
- Tested assignee avatar.

Reviewed at https://reviews.reviewboard.org/r/7578/