Swap out Pikaday for bootstrap-datepicker.
I had used Pikaday in the calendar UI, but later on in the all status reports
list, I used bootstrap-datepicker, which fits in slightly better with the look
and feel (and has an equally OK API). This changes out the one in the calendar
and removes the dependency.

Testing done:
Edited and saved some dates in the calendar UI.

Reviewed at https://reviews.reviewboard.org/r/7579/