Swap out tagsinput for selectize.
The bootstrap-tagsinput widget is kind of janky. This change swaps it out for
selectize. As part of this, I've implemented an x-editable backend for it. This
is a little annoying because selectize's setValue() method doesn't seem to work
right, so I have to re-init it when rendering the value, but looking through
other x-editable backends, it looks like that's not uncommon.

The existing tagsinput widget was currently only used in the calendar editor.
As part of this, I've made it so the groups editors on the user detail and all
status report views use it, too.

Testing done:
Tested groups editors on all views that have them.

Reviewed at https://reviews.reviewboard.org/r/7581/