Add an "All Students List" for mentors to use.
This change adds a simple new view that just lists all of the registered users
and links to the detail pages. This will make it possible to use the sonar UI
to set up new students to be in the right groups and have school/email/etc
data, instead of having to edit it in MongoDB directly.

Along with this, I've fixed it up so only mentors see the list of students in
the sidebar.

Testing done:
Looked at the new user list. Checked that both it and the sidebar linked to the
detail views.

Reviewed at https://reviews.reviewboard.org/r/7582/