Improve the design and usability of the projects list.
This change makes it so that the main body of the projects list and the table
of contents scroll independently. Additionally, scrolling the body will
auto-highlight the matching item in the table of contents, which makes it much
easier to figure out where you are.

Testing done:
Scrolled around a whole bunch. Made sure that everything updated correctly, and
that links scrolled to the right place on the page.

Reviewed at https://reviews.reviewboard.org/r/7586/