Implement scrolling behavior for projects list nav.
When scrolling the main body of the projects list, it's nice to also scroll the
nav to show the highlighted item. This implements that.

Testing done:
Scrolled around with a variety of window sizes and scroll positions. Saw that
everything worked as expected.

Reviewed at https://reviews.reviewboard.org/r/7587/