List projects on user detail page.
This change adds a "Projects" section that has a list of all of the assigned
projects for each student. This is populated from the project assignments in
Asana.

These do not yet link to the project description (since there's some issue with
scrolling to a project when first opening the page), but it's pretty good for
now.

Testing done:
Looked at all of our students and saw the project lists.

Reviewed at https://reviews.reviewboard.org/r/7712/