Add a "notes" field to the user detail view.
This adds a new field for us to use for keeping student notes directly in
Sonar. It's not quite like Google Docs in that we have to be wary about
conflicts with multiple editors, but it'll be nice to have it in one place.

Testing done:
Used this to enter in the shirt sizes.

Reviewed at https://reviews.reviewboard.org/r/7848/