Create extension for auto-assigning students to the students group
A simple extension for auto-assigning students working on Review Board to
the "students" group.

I can imagine this maybe being useful for other projects as well, so I can
see this expanding out to allow multiple auto-assignments, but for now I've
just hard-coded the "students" group.

Testing done:
Created some test users and a "students" group. Entered the extension
configuration page and added some users in the comma-separated list to
auto-assign. Created review requests for the users in the list and saw
that they were correctly added to the "students" group. The review requests
from the other users were not.

Also verified that the "open" counts in the dashboard for the group
was updated correctly after creating an auto-assigned review request.

Reviewed at https://reviews.reviewboard.org/r/7872/