Add code reviews to sonar.
This change adds links to code reviews inside sonar. This depends on /r/7888 to
actually get the URL to the code review, and there's currently no good text to
put in the link, but it will at least give us a list of links we can see and
copy/paste.

This also adds a spinner to the review request list while it's loading.

Testing done:
Looked at a page that had both review requests and reviews.

Reviewed at https://reviews.reviewboard.org/r/7889/