Refactor student projects list a bit, link to user detail views from project.
This change moves the code that renders an individual student project out into
a new view, which I'll eventually be making use of for a view of an individual
project (for use when linking from the user detail view).

This also makes it so users who have permission can click the user avatar to
jump to the user detail view for that user.

Testing done:
Looked through the project list and saw that everything still worked. Clicked
on the user avatar and saw that it jumped to the right user. Set myself to be a
student and saw that the users were no longer links.

Reviewed at https://reviews.reviewboard.org/r/7935/