Switch to access tokens for Asana authentication.
Asana is deprecating their old API key authentication in favor of a new system
of access tokens (or OAuth, which is overkill for our needs). They'll be
turning it off next month, so we need to switch over. Luckily, because they're
the authors of the library we use for accessing the API, it's super easy.

Testing done:
Ran the server and opened the project ideas page. Saw that it correctly fetched
everything.

Reviewed at https://reviews.reviewboard.org/r/8551/