Move over to webpack.
JSPM is cool, but System.js does some weird stuff which is very hard to
maintain, and even harder to work in new dependencies for neat stuff we
might want to use. Webpack is a little less magic but a little more
dependable, and a lot faster to deploy (since we don't have to reinstall
the world).

This change moves us over, and does various updates to the frontend code
to make everything work. There are a couple small bugs which seem to
just be some of our dependencies (like x-editable) being unmaintained
and getting fairly long in the tooth, but those bugs are not visible to
students and don't impact actually using sonar.

Testing done:
Tested the various features of sonar, for both mentors and students.
Everything seems to be pretty functional.