Convert header to react.
This change converts the header to use react internally. It's still
wrapped by a Backbone view, but that will change in the next commit,
which moves the main application frame to use react.

Testing done:
Verified that the header functioned correctly, including the toggle for
the "manage" state.

Reviewed at https://reviews.reviewboard.org/r/9558/