Convert the rest of the status report views to React.
This change contains React conversions for the the status report viewer
(which just renders a submitted status report) and the status report
editor (which is the big one).

As part of this, I'm moving away from EpicEditor. When I first created
Sonar, this was an ok package, but it's always had some bugs, and it
hasn't had any real updates in four years. Instead, I've integrated a
new package called "react-mde", which is a fully controlled react
component for editing and previewing markdown.

I've also tweaked the UI a bit in the status report editor, moving the
save button into the header line, as well as creating a new "Preview"
toggle switch (which replaces the one that EpicEditor would show inside
its edit region).

Testing done:
Viewed, listed, and edited a bunch of status reports. Verified that
editing worked whether the status report was brand new (in which case it
was pre-populated with the template) or existing (in which case it was
properly loaded).

Reviewed at https://reviews.reviewboard.org/r/9584/