Convert the user detail view to react.
Here's the big one. This change converts the entire user detail view to
react, meaning the only backbone view that's left is the main
application view shell. Because this is the most complex page in the
entire application, it's a somewhat large change, but hopefully nothing
in here is too surprising.

Testing done:
- Verified that viewing and editing data in all sections of the user
  detail view worked correctly.
- Switched between a bunch of different users and saw the correct data
  for each of them, even though it's no longer replacing the entire
  view.

Reviewed at https://reviews.reviewboard.org/r/9675/