Start adding Redux for data management.
Backbone models and collections are not a great way to store data for a
react application. This change starts the (probably lengthy) process of
migrating towards a Redux data management model. The first thing which
is really migrated here is the project ideas, which is one of the most
isolated pieces of data.

In order to match things up here, this also includes the start of
migration for the users and the "manage" state. Those are only very
partially migrated, so there's a lot more work to do there.

Testing done:
Navigated around through the three views that use projects (the project
list, the single project view, and the user detail view) and saw that
everything worked as expected.

Reviewed at https://reviews.reviewboard.org/r/9678/