Fix load-slack-users for new imports.
I never updated the load-slack-users script for the new ES6 modules
scheme for everything, which explains why we were slowly losing the
realnames (due to a bug with the `real_name` field that we were using).
While I was doing this I changed things to use the
`profile.real_name_normalized` field, which should hopefully not exhibit
the same problem.